requirements.txt
# AI / LLM
langgraph>=0.2
langchain-openai>=0.2
openai>=1.0

# Web UI
streamlit>=1.38

# Data / DB
pandas>=2.0
pyodbc>=5.0
pyyaml>=6.0

# Templating / validation
jinja2>=3.1
pydantic>=2.0
pydantic-settings>=2.0

# Utilities
python-dotenv>=1.0
python-dateutil>=2.8

# Testing
pytest>=8.0
